+++ test/core/InvertedTakerOperations.t.sol
@@ -52,7 +52,7 @@ contract InvertedTakerOperationsTest is ITaker, MangroveTest {
     mkr.newOffer(0.1 ether, 0.1 ether, 100_000, 0);
     _takerTrade = checkPay;
     toPay = 0.2 ether;
-    (, uint gave,,) = mgv.marketOrder($(base), $(quote), 0.2 ether, 0.2 ether, true);
+    (, uint gave,,) = mgv.marketOrder($(base), $(quote), uint(0.2 ether), 0.2 ether, true);
     assertEq(quoteBalance - gave, quote.balanceOf($(this)), "totalGave should be sum of taker flashborrows");
   }
 
@@ -67,7 +67,7 @@ contract InvertedTakerOperationsTest is ITaker, MangroveTest {
     uint _ofr = mkr.newOffer(0.1 ether, 0.1 ether, 100_000, 0);
     _takerTrade = revertTrade;
     skipCheck = true;
-    try mgv.marketOrder($(base), $(quote), 0.2 ether, 0.2 ether, true) {
+    try mgv.marketOrder($(base), $(quote), uint(0.2 ether), 0.2 ether, true) {
       fail("Market order should have reverted");
     } catch Error(string memory reason) {
       assertEq(REVERT_TRADE_REASON, reason, "Unexpected throw");
@@ -87,7 +87,7 @@ contract InvertedTakerOperationsTest is ITaker, MangroveTest {
   function test_taker_refuses_to_deliver_during_trade() public {
     mkr.newOffer(0.1 ether, 0.1 ether, 100_000, 0);
     _takerTrade = refusePayTrade;
-    try mgv.marketOrder($(base), $(quote), 0.2 ether, 0.2 ether, true) {
+    try mgv.marketOrder($(base), $(quote), uint(0.2 ether), 0.2 ether, true) {
       fail("Market order should have reverted");
     } catch Error(string memory reason) {
       assertEq(reason, "mgv/takerFailToPayTotal", "Unexpected throw message");
@@ -125,7 +125,7 @@ contract InvertedTakerOperationsTest is ITaker, MangroveTest {
     emit OfferSuccess($(base), $(quote), 1, $(this), 0.1 ether, 0.1 ether);
     expectFrom($(mgv));
     emit OfferSuccess($(base), $(quote), 2, $(this), 0.1 ether, 0.1 ether);
-    (uint got, uint gave,,) = mgv.marketOrder($(base), $(quote), 0.1 ether, 0.1 ether, true);
+    (uint got, uint gave,,) = mgv.marketOrder($(base), $(quote), uint(0.1 ether), 0.1 ether, true);
     assertEq(quoteBalance - gave - 0.1 ether, quote.balanceOf($(this)), "Incorrect transfer (gave) during reentrancy");
     assertEq(baseBalance + got + 0.1 ether, base.balanceOf($(this)), "Incorrect transfer (got) during reentrancy");
   }
